refactor: use global session store
